Live Kafka test.
