+++ server/swimos_connector_kafka/Cargo.toml
@@ -34,7 +34,7 @@ regex = { workspace = true }
 frunk = { workspace = true }
 
 [dev-dependencies]
-tokio = { workspace = true, features = ["rt-multi-thread", "signal"] }
+tokio = { workspace = true, features = ["rt"] }
 uuid = { workspace = true, features = ["v4"] }
 bytes = { workspace = true }
 parking_lot = { workspace = true }
